+++ polaris.shopify.com/next.config.js
@@ -192,11 +192,6 @@ const deprecated = [
     permanent: false,
   },
   {
-    source: '/components/grid',
-    destination: '/components/deprecated/grid',
-    permanent: false,
-  },
-  {
     source: '/components/heading',
     destination: '/components/deprecated/heading',
     permanent: false,
@@ -388,6 +383,11 @@ const layoutAndStructure = [
   {
     source: '/components/forms/form-layout',
     destination: '/components/layout-and-structure/form-layout',
+    permanent: false,
+  },
+  {
+    source: '/components/grid',
+    destination: '/components/layout-and-structure/grid',
     permanent: false,
   },
   {