+++ polaris-codemods/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shopify/polaris-codemods
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`247f5eea8`](https://github.com/Shopify/polaris/commit/247f5eea859b3ab348dcb18c568f18d8d859140e)]:
+  - @shopify/polaris-tokens@6.10.0
+  - @shopify/stylelint-polaris@9.0.3
+
 ## 0.0.1
 
 ### Patch Changes