+++ polaris-migrator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shopify/polaris-migrator
 
+## 0.17.1
+
+### Patch Changes
+
+- Updated dependencies [[`247f5eea8`](https://github.com/Shopify/polaris/commit/247f5eea859b3ab348dcb18c568f18d8d859140e)]:
+  - @shopify/polaris-tokens@6.10.0
+  - @shopify/stylelint-polaris@9.0.3
+
 ## 0.17.0
 
 ### Minor Changes